requirements.txt
selenium>=4.16.0
pytest>=8.0.0
google-generativeai>=0.3.0
faiss-cpu>=1.7.4
numpy>=1.24.0
lxml>=4.9.0
python-dotenv>=1.0.0
webdriver-manager>=4.0.0
pytest-html>=4.1.0
pytest-xdist>=3.5.0

